backend/requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
mysql-connector-python==8.1.0
python-dotenv==1.0.0
email-validator==2.1.0.post1